refactor: 重构事件处理器模块 - 分离处理器实现
- 将事件处理器从 concrete_handlers.py 拆分到独立模块
- 创建 handlers/ 目录包含各类事件处理器
- 改进类型提示和文档字符串
- 修正分支名称拼写错误 (splie -> split)
