+++ .gitignore
@@ -33,6 +33,7 @@ node_modules/
 CLAUDE.md
 .mcp.json
 QWEN.md
+AGENTS.md
 # Ai学习用的文件
 knowledge_base*
 # 运行时生成的文件
